move most components into internal/core
in this way coverage can be computed correctly.
