+++ internal/core/core_test.go
@@ -1,0 +1,216 @@
+package core
+
+import (
+	"io/ioutil"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"strconv"
+	"testing"
+	"time"
+
+	"github.com/stretchr/testify/require"
+)
+
+type container struct {
+	name string
+}
+
+func newContainer(image string, name string, args []string) (*container, error) {
+	c := &container{
+		name: name,
+	}
+
+	exec.Command("docker", "kill", "rtsp-simple-server-test-"+name).Run()
+	exec.Command("docker", "wait", "rtsp-simple-server-test-"+name).Run()
+
+	// --network=host is needed to test multicast
+	cmd := []string{
+		"docker", "run",
+		"--network=host",
+		"--name=rtsp-simple-server-test-" + name,
+		"rtsp-simple-server-test-" + image,
+	}
+	cmd = append(cmd, args...)
+	ecmd := exec.Command(cmd[0], cmd[1:]...)
+	ecmd.Stdout = nil
+	ecmd.Stderr = os.Stderr
+
+	err := ecmd.Start()
+	if err != nil {
+		return nil, err
+	}
+
+	time.Sleep(1 * time.Second)
+
+	return c, nil
+}
+
+func (c *container) close() {
+	exec.Command("docker", "kill", "rtsp-simple-server-test-"+c.name).Run()
+	exec.Command("docker", "wait", "rtsp-simple-server-test-"+c.name).Run()
+	exec.Command("docker", "rm", "rtsp-simple-server-test-"+c.name).Run()
+}
+
+func (c *container) wait() int {
+	exec.Command("docker", "wait", "rtsp-simple-server-test-"+c.name).Run()
+	out, _ := exec.Command("docker", "inspect", "rtsp-simple-server-test-"+c.name,
+		"-f", "{{.State.ExitCode}}").Output()
+	code, _ := strconv.ParseInt(string(out[:len(out)-1]), 10, 64)
+	return int(code)
+}
+
+func (c *container) ip() string {
+	out, _ := exec.Command("docker", "inspect", "rtsp-simple-server-test-"+c.name,
+		"-f", "{{range .NetworkSettings.Networks}}{{.IPAddress}}{{end}}").Output()
+	return string(out[:len(out)-1])
+}
+
+func writeTempFile(byts []byte) (string, error) {
+	tmpf, err := ioutil.TempFile(os.TempDir(), "rtsp-")
+	if err != nil {
+		return "", err
+	}
+	defer tmpf.Close()
+
+	_, err = tmpf.Write(byts)
+	if err != nil {
+		return "", err
+	}
+
+	return tmpf.Name(), nil
+}
+
+func newInstance(conf string) (*Core, bool) {
+	if conf == "" {
+		return New([]string{})
+	}
+
+	tmpf, err := writeTempFile([]byte(conf))
+	if err != nil {
+		return nil, false
+	}
+	defer os.Remove(tmpf)
+
+	return New([]string{tmpf})
+}
+
+var serverCert = []byte(`-----BEGIN CERTIFICATE-----
+MIIDazCCAlOgAwIBAgIUXw1hEC3LFpTsllv7D3ARJyEq7sIwDQYJKoZIhvcNAQEL
+BQAwRTELMAkGA1UEBhMCQVUxEzARBgNVBAgMClNvbWUtU3RhdGUxITAfBgNVBAoM
+GEludGVybmV0IFdpZGdpdHMgUHR5IEx0ZDAeFw0yMDEyMTMxNzQ0NThaFw0zMDEy
+MTExNzQ0NThaMEUxCzAJBgNVBAYTAkFVMRMwEQYDVQQIDApTb21lLVN0YXRlMSEw
+HwYDVQQKDBhJbnRlcm5ldCBXaWRnaXRzIFB0eSBMdGQwggEiMA0GCSqGSIb3DQEB
+AQUAA4IBDwAwggEKAoIBAQDG8DyyS51810GsGwgWr5rjJK7OE1kTTLSNEEKax8Bj
+zOyiaz8rA2JGl2VUEpi2UjDr9Cm7nd+YIEVs91IIBOb7LGqObBh1kGF3u5aZxLkv
+NJE+HrLVvUhaDobK2NU+Wibqc/EI3DfUkt1rSINvv9flwTFu1qHeuLWhoySzDKEp
+OzYxpFhwjVSokZIjT4Red3OtFz7gl2E6OAWe2qoh5CwLYVdMWtKR0Xuw3BkDPk9I
+qkQKx3fqv97LPEzhyZYjDT5WvGrgZ1WDAN3booxXF3oA1H3GHQc4m/vcLatOtb8e
+nI59gMQLEbnp08cl873bAuNuM95EZieXTHNbwUnq5iybAgMBAAGjUzBRMB0GA1Ud
+DgQWBBQBKhJh8eWu0a4au9X/2fKhkFX2vjAfBgNVHSMEGDAWgBQBKhJh8eWu0a4a
+u9X/2fKhkFX2vjAPBgNVHRMBAf8EBTADAQH/MA0GCSqGSIb3DQEBCwUAA4IBAQBj
+3aCW0YPKukYgVK9cwN0IbVy/D0C1UPT4nupJcy/E0iC7MXPZ9D/SZxYQoAkdptdO
+xfI+RXkpQZLdODNx9uvV+cHyZHZyjtE5ENu/i5Rer2cWI/mSLZm5lUQyx+0KZ2Yu
+tEI1bsebDK30msa8QSTn0WidW9XhFnl3gRi4wRdimcQapOWYVs7ih+nAlSvng7NI
+XpAyRs8PIEbpDDBMWnldrX4TP6EWYUi49gCp8OUDRREKX3l6Ls1vZ02F34yHIt/7
+7IV/XSKG096bhW+icKBWV0IpcEsgTzPK1J1hMxgjhzIMxGboAeUU+kidthOob6Sd
+XQxaORfgM//NzX9LhUPk
+-----END CERTIFICATE-----
+`)
+
+var serverKey = []byte(`-----BEGIN RSA PRIVATE KEY-----
+MIIEogIBAAKCAQEAxvA8skudfNdBrBsIFq+a4ySuzhNZE0y0jRBCmsfAY8zsoms/
+KwNiRpdlVBKYtlIw6/Qpu53fmCBFbPdSCATm+yxqjmwYdZBhd7uWmcS5LzSRPh6y
+1b1IWg6GytjVPlom6nPxCNw31JLda0iDb7/X5cExbtah3ri1oaMkswyhKTs2MaRY
+cI1UqJGSI0+EXndzrRc+4JdhOjgFntqqIeQsC2FXTFrSkdF7sNwZAz5PSKpECsd3
+6r/eyzxM4cmWIw0+Vrxq4GdVgwDd26KMVxd6ANR9xh0HOJv73C2rTrW/HpyOfYDE
+CxG56dPHJfO92wLjbjPeRGYnl0xzW8FJ6uYsmwIDAQABAoIBACi0BKcyQ3HElSJC
+kaAao+Uvnzh4yvPg8Nwf5JDIp/uDdTMyIEWLtrLczRWrjGVZYbsVROinP5VfnPTT
+kYwkfKINj2u+gC6lsNuPnRuvHXikF8eO/mYvCTur1zZvsQnF5kp4GGwIqr+qoPUP
+bB0UMndG1PdpoMryHe+JcrvTrLHDmCeH10TqOwMsQMLHYLkowvxwJWsmTY7/Qr5S
+Wm3PPpOcW2i0uyPVuyuv4yD1368fqnqJ8QFsQp1K6QtYsNnJ71Hut1/IoxK/e6hj
+5Z+byKtHVtmcLnABuoOT7BhleJNFBksX9sh83jid4tMBgci+zXNeGmgqo2EmaWAb
+agQslkECgYEA8B1rzjOHVQx/vwSzDa4XOrpoHQRfyElrGNz9JVBvnoC7AorezBXQ
+M9WTHQIFTGMjzD8pb+YJGi3gj93VN51r0SmJRxBaBRh1ZZI9kFiFzngYev8POgD3
+ygmlS3kTHCNxCK/CJkB+/jMBgtPj5ygDpCWVcTSuWlQFphePkW7jaaECgYEA1Blz
+ulqgAyJHZaqgcbcCsI2q6m527hVr9pjzNjIVmkwu38yS9RTCgdlbEVVDnS0hoifl
++jVMEGXjF3xjyMvL50BKbQUH+KAa+V4n1WGlnZOxX9TMny8MBjEuSX2+362vQ3BX
+4vOlX00gvoc+sY+lrzvfx/OdPCHQGVYzoKCxhLsCgYA07HcviuIAV/HsO2/vyvhp
+xF5gTu+BqNUHNOZDDDid+ge+Jre2yfQLCL8VPLXIQW3Jff53IH/PGl+NtjphuLvj
+7UDJvgvpZZuymIojP6+2c3gJ3CASC9aR3JBnUzdoE1O9s2eaoMqc4scpe+SWtZYf
+3vzSZ+cqF6zrD/Rf/M35IQKBgHTU4E6ShPm09CcoaeC5sp2WK8OevZw/6IyZi78a
+r5Oiy18zzO97U/k6xVMy6F+38ILl/2Rn31JZDVJujniY6eSkIVsUHmPxrWoXV1HO
+y++U32uuSFiXDcSLarfIsE992MEJLSAynbF1Rsgsr3gXbGiuToJRyxbIeVy7gwzD
+94TpAoGAY4/PejWQj9psZfAhyk5dRGra++gYRQ/gK1IIc1g+Dd2/BxbT/RHr05GK
+6vwrfjsoRyMWteC1SsNs/CurjfQ/jqCfHNP5XPvxgd5Ec8sRJIiV7V5RTuWJsPu1
++3K6cnKEyg+0ekYmLertRFIY6SwWmY1fyKgTvxudMcsBY7dC4xs=
+-----END RSA PRIVATE KEY-----
+`)
+
+func TestCoreHotReloading(t *testing.T) {
+	confPath := filepath.Join(os.TempDir(), "rtsp-conf")
+
+	err := ioutil.WriteFile(confPath, []byte("paths:\n"+
+		"  test1:\n"+
+		"    runOnDemand: ffmpeg -hide_banner -loglevel error -re -i ../../testimages/ffmpeg/emptyvideo.mkv -c copy -f rtsp rtsp://localhost:$RTSP_PORT/$RTSP_PATH\n"+
+		"  test3:\n"+
+		"    runOnInit: echo aaa\n"+
+		"  test4:\n"+
+		"    runOnInit: echo bbb\n"),
+		0o644)
+	require.NoError(t, err)
+	defer os.Remove(confPath)
+
+	p, ok := New([]string{confPath})
+	require.Equal(t, true, ok)
+	defer p.close()
+
+	time.Sleep(1 * time.Second)
+
+	func() {
+		cnt1, err := newContainer("ffmpeg", "dest", []string{
+			"-i", "rtsp://localhost:8554/test1",
+			"-vframes", "1",
+			"-f", "image2",
+			"-y", "/dev/null",
+		})
+		require.NoError(t, err)
+		defer cnt1.close()
+		require.Equal(t, 0, cnt1.wait())
+	}()
+
+	err = ioutil.WriteFile(confPath, []byte("paths:\n"+
+		"  test2:\n"+
+		"    runOnDemand: ffmpeg -hide_banner -loglevel error -re -i ../../testimages/ffmpeg/emptyvideo.mkv -c copy -f rtsp rtsp://localhost:$RTSP_PORT/$RTSP_PATH\n"+
+		"  test3:\n"+
+		"  test4:\n"+
+		"    runOnInit: echo bbb\n"),
+		0o644)
+	require.NoError(t, err)
+
+	time.Sleep(1 * time.Second)
+
+	func() {
+		cnt1, err := newContainer("ffmpeg", "dest", []string{
+			"-i", "rtsp://localhost:8554/test1",
+			"-vframes", "1",
+			"-f", "image2",
+			"-y", "/dev/null",
+		})
+		require.NoError(t, err)
+		defer cnt1.close()
+		require.Equal(t, 1, cnt1.wait())
+	}()
+
+	func() {
+		cnt1, err := newContainer("ffmpeg", "dest", []string{
+			"-i", "rtsp://localhost:8554/test2",
+			"-vframes", "1",
+			"-f", "image2",
+			"-y", "/dev/null",
+		})
+		require.NoError(t, err)
+		defer cnt1.close()
+		require.Equal(t, 0, cnt1.wait())
+	}()
+}
