+++ internal/rtmp/client.go
@@ -11,7 +11,7 @@ import (
 
 // DialContext connects to a server in reading mode.
 func DialContext(ctx context.Context, address string) (*Conn, error) {
-	// https://github.com/aler9/rtmp/blob/master/format/rtmp/readpublisher.go#L74
+	// https://github.com/aler9/rtmp/blob/3be4a55359274dcd88762e72aa0a702e2d8ba2fd/format/rtmp/client.go#L74
 
 	u, err := url.Parse(address)
 	if err != nil {
